Update 100-elf_header.c
Nothing

+++ 0x15-file_io/100-elf_header.c
@@ -309,7 +309,7 @@ void printingEntry(unsigned long int e_entry, unsigned char *e_ident)
 		e_entry = (e_entry << 16) | (e_entry >> 16);
 	}
 	if (e_ident[EI_CLASS] == ELFCLASS32)
-		printf("%#x\n", (unsigned long)e_entry);
+		printf("%#x\n", (unsigned int)e_entry);
 	else
 		printf("%#lx\n", e_entry);
 }
